+++ src/qnethernet/drivers/driver_teensy41.c
@@ -931,7 +931,7 @@ FLASHMEM bool driver_init(void) {
 
   ENET_RCR = 0
              | ENET_RCR_NLC        // Payload length is checked
-             | ENET_RCR_MAX_FL(MAX_FRAME_LEN)
+             | ENET_RCR_MAX_FL((MAX_FRAME_LEN) + 4)
              | ENET_RCR_CFEN       // Discard non-pause MAC control frames
              | ENET_RCR_CRCFWD     // CRC is stripped (ignored if PADEN)
              | ENET_RCR_PADEN      // Padding is removed