+++ src/qnethernet/drivers/driver_teensy41.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #define MTU           1500
-#define MAX_FRAME_LEN 1522  /* Includes the 4-byte FCS (frame check sequence) */
+#define MAX_FRAME_LEN 1518  /* Does not include the 4-byte FCS (frame check sequence) */
 
 // lwIP options