+++ src/qnethernet/drivers/driver_unsupported.h
@@ -7,4 +7,4 @@
 #pragma once
 
 #define MTU           1500
-#define MAX_FRAME_LEN 1522
+#define MAX_FRAME_LEN 1518  /* Does not include the 4-byte FCS (frame check sequence) */