+++ test/test_ethernet/test_main.cpp
@@ -1541,7 +1541,7 @@ static void test_other_state() {
   TEST_ASSERT_EQUAL_MESSAGE(enet_get_mtu(), Ethernet.mtu(), "Expected default MTU");
   TEST_ASSERT_EQUAL_MESSAGE(enet_get_max_frame_len(), EthernetFrame.maxFrameLen(),
                             "Expected default max. frame len");
-  TEST_ASSERT_EQUAL_MESSAGE(64, EthernetFrame.minFrameLen(), "Expected default min. frame len");
+  TEST_ASSERT_EQUAL_MESSAGE(60, EthernetFrame.minFrameLen(), "Expected default min. frame len");
   TEST_ASSERT_EQUAL_MESSAGE(MDNS_MAX_SERVICES, MDNS.maxServices(), "Expected default mDNS max. services");
 }